Refactor: Implement platform specific IO and UDP
