+++ pubspec.yaml
@@ -56,6 +56,7 @@ dependencies:
   latlong2: ^0.9.0
   auto_size_text: ^3.0.0
   intl: any
+  universal_io: ^2.2.2
   flutter_launcher_icons: ^0.14.4
   introduction_screen: ^3.1.12
   cached_network_image: ^3.3.1
